GRAD2-1850 Year-End Disribution run (DISTRUN_YE) - reports School-Level (DISTREP_YE_SC) and District-Level (DISTREP_YE_SD)

+++ api/src/main/java/ca/bc/gov/educ/grad/report/dto/reports/bundle/decorator/SchoolReportDecorator.java
@@ -61,7 +61,7 @@ public class SchoolReportDecorator extends DocumentBundleDecorator {
      */
     @Override
     protected NumberLabel createPageCountLabel(int count) {
-        NumberLabel label = new PageNumberLabel(new Point2D.Float(490, 810), count, 0);
+        NumberLabel label = new PageNumberLabel(new Point2D.Float(495, 810), count, 0);
         label.setNumberFormat("Page: %s%s");
         label.setLabelPrefix("");
         return label;
